backend/requirements.txt
# SQL Agent Requirements
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0
pandas>=1.5.0
requests>=2.31.0
matplotlib>=3.7.0
numpy>=1.24.0

# Flask Backend Requirements
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0